Fix: enforce integer validation for year in AdminRequestSchema

+++ packages/types/src/requests/admin/getAllRequests/schemas.ts
@@ -6,7 +6,7 @@ export const AdminRequestSchema = z.object({
   id: IdSchema.describe("The ID of the request"),
   organizationName: z.string().min(1).describe("The name of the organization"),
   type: z.enum(SubmissionSubjectType).describe("The type of the request"),
-  year: z.number().describe("The year of the request"),
+  year: z.number().int().describe("The year of the request"),
   status: z.enum(SubmissionStatus).describe("The status of the request"),
   requestedAt: z.iso.datetime().describe("The date the request was made"),
 });
